+++ R/generics_easystats.R
@@ -18,7 +18,7 @@
 #' mod <- lm(mpg ~ wt + cyl, data = mtcars)
 #' broomExtra::tidy_parameters(mod)
 #' @importFrom rlang is_null
-#' @importFrom parameters model_parameters
+#' @importFrom parameters model_parameters standardize_names
 #'
 #' @export
 
@@ -26,14 +26,14 @@ tidy_parameters <- function(x, conf.int = TRUE, ...) {
   # easystats family ---------------------------------------
   # check if `easystats` family has a tidy method for a given object
   m <- tryCatch(
-    expr = easystats_to_tidy_names(parameters::model_parameters(x, ...)),
+    expr = standardize_names(parameters::model_parameters(x, ...), style = "broom"),
     error = function(e) NULL
   )
 
 
   if (rlang::is_null(m)) {
     m <- tryCatch(
-      expr = easystats_to_tidy_names(parameters::model_parameters(x)),
+      expr = standardize_names(parameters::model_parameters(x), style = "broom"),
       error = function(e) NULL
     )
   }
@@ -59,7 +59,7 @@ tidy_parameters <- function(x, conf.int = TRUE, ...) {
   if (rlang::is_null(m)) m <- tryCatch(as_tibble(x, ...), error = function(e) NULL)
 
   # return the final object
-  return(m)
+  if (rlang::is_null(m)) m else as_tibble(m)
 }
 
 
@@ -95,7 +95,9 @@ glance_performance <- function(x, ...) {
   # easystats family ---------------------------------------
   # check if `easystats` family has a tidy method for a given object
   df_performance <- tryCatch(
-    expr = easystats_to_tidy_names(performance::model_performance(x, metrics = "all", ...)),
+    expr = standardize_names(performance::model_performance(x, metrics = "all", ...),
+      style = "broom"
+    ),
     error = function(e) NULL
   )
 
@@ -114,5 +116,5 @@ glance_performance <- function(x, ...) {
   if (rlang::is_null(df_performance)) df_combined <- df_broom
 
   # return the final object
-  return(df_combined)
+  if (rlang::is_null(df_combined)) df_combined else as_tibble(df_combined)
 }