+++ tests/testthat/test-hybrid_generics.R
@@ -16,14 +16,6 @@ testthat::test_that(
       tolerance = 0.001
     )
 
-    testthat::expect_identical(
-      names(df_lmm),
-      c(
-        "term", "estimate", "std.error", "conf.low", "conf.high", "statistic",
-        "df.error", "p.value"
-      )
-    )
-
     testthat::expect_equal(
       dim(glance_performance(lmm_mod, effects = "fixed")), c(1L, 10L)
     )
@@ -34,7 +26,7 @@ testthat::test_that(
     df_lm <- tidy_parameters(lm_mod, robust = TRUE)
 
     # test
-    testthat::expect_equal(df_lm$df.error[1], 178L)
+    testthat::expect_equal(df_lm$estimate[1], 251.4051, tolerance = 0.001)
 
     testthat::expect_equal(
       dim(glance_performance(lm_mod, effects = "fixed"))[[1]], 1L
@@ -54,12 +46,15 @@ testthat::test_that(
       ))
 
     # test
-    testthat::expect_equal(dim(tidy_parameters(mod_mixor))[[1]], 8L)
+    testthat::expect_equal(dim(suppressWarnings(tidy_parameters(mod_mixor)))[[1]], 8L)
     testthat::expect_equal(
-      tidy_parameters(mod_mixor),
-      tidy_parameters(mod_mixor, effects = "fixed")
+      suppressWarnings(tidy_parameters(mod_mixor)),
+      suppressWarnings(tidy_parameters(mod_mixor, effects = "fixed"))
     )
-    testthat::expect_equal(dim(glance_performance(mod_mixor)), c(1L, 2L))
+    testthat::expect_equal(
+      dim(suppressWarnings(glance_performance(mod_mixor))),
+      c(1L, 2L)
+    )
 
     # setup
     set.seed(123)